# 05_Application_Design/05_Application_Design/API_Specification.yaml
openapi: 3.0.3
info:
  title: Finished Goods Warehouse API Wrapper
  description: API wrapper gọi SQL Server Stored Procedure cho hệ thống WMS kho thành phẩm. Business logic nằm trong Stored Procedure.
  version: 0.6.0
servers:
  - url: https://wms-api.local/api/v1
    description: Internal API wrapper
  - url: http://localhost:5000/api/v1
    description: Local development
...
paths:
...
  /stock-type/change:
    post:
      tags: [StockType]
      summary: Chuyển stock type, ví dụ BLOCKED cho dư đơn hoặc vấn đề chất lượng
      description: Wrapper gọi Stored Procedure usp_StockType_Change_Post.
      operationId: changeStockType
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockTypeChangeRequest'
      responses:
        '200':
          description: Kết quả xử lý từ Stored Procedure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
...
components:
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: true
      schema:
        type: string
      description: Mã chống gửi lặp, được truyền xuống Stored Procedure.
  schemas:
    CommandResult:
      type: object
      properties:
        status:
          type: string
          enum: [SUCCESS, FAILED, DUPLICATED, CONFLICT, UNAUTHORIZED]
        message:
          type: string
        error_code:
          type: string
          nullable: true
        document_no:
          type: string
          nullable: true
        object_code:
          type: string
          nullable: true
        request_id:
          type: string
        trace_id:
          type: string
      required: [status, message, request_id]
...
    StockTypeChangeRequest:
      type: object
      properties:
        object_type:
          type: string
          enum: [THUNG60, PACK360, PALLET, LOCATION]
        object_codes:
          type: array
          items:
            type: string
        new_stock_type:
          type: string
          enum: [UNRESTRICTED, BLOCKED, RETURNED, TEMPORARY_ISSUE, CUSTOMER_OWNED, SCRAP]
        reason_code:
          type: string
          enum: [OEM_SURPLUS, QUALITY_ISSUE, PARTIAL_REMAINING, DATA_EXCEPTION, WAITING_DECISION]
        note:
          type: string
        user_code:
          type: string
      required: [object_type, object_codes, new_stock_type, reason_code, user_code]
...